+++ unoplat-code-confluence-ingestion/code-confluence-flow-bridge/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.62.1](https://github.com/unoplat/unoplat-code-confluence/compare/code-confluence-flow-bridge-v0.62.0...code-confluence-flow-bridge-v0.62.1) (2025-11-22)
+
+
+### Bug Fixes
+
+* **graph-db:** Update CodeConfluenceGitRepository node creation to not cause cypher merge conflicts ([bde1103](https://github.com/unoplat/unoplat-code-confluence/commit/bde1103e6a744a6c3024e204bc5b5393232c0f30))
+* **graph-db:** Update CodeConfluenceGitRepository node creation to not cause cypher merge conflicts ([dd15257](https://github.com/unoplat/unoplat-code-confluence/commit/dd1525719cd5d4b7ab07528a0621023cd935d6e9))
+
+## [0.62.0](https://github.com/unoplat/unoplat-code-confluence/compare/code-confluence-flow-bridge-v0.61.1...code-confluence-flow-bridge-v0.62.0) (2025-11-21)
+
+
+### Features
+
+* support more repository providers starting with github enterprise. Improved Frontend Architecture and onboarding and some bugs squashed ([0974c76](https://github.com/unoplat/unoplat-code-confluence/commit/0974c76fe9df68b86fb1b180db1b41e04a16c9dd))
+
 ## [0.61.1](https://github.com/unoplat/unoplat-code-confluence/compare/code-confluence-flow-bridge-v0.61.0...code-confluence-flow-bridge-v0.61.1) (2025-11-21)